Added reference count Garbage collector

+++ Heap-Objects/Dynamic-Array/dynamic_array.h
@@ -1,7 +1,8 @@
 #ifndef DYNAMIC_ARRAY_H
 #define DYNAMIC_ARRAY_H
-#include <stdlib.h>
-#include "../Primitve-Objects/object.h"
+#include <stddef.h>
+#include "Heap-Objects/Primitive-Objects/object.h"
+#include "Garbage-Collector/Reference-Counting/ref_counter.h"
 
 typedef struct Object obj_t;
 
@@ -12,8 +13,8 @@ typedef struct DynamicArray
     obj_t **objects; // pointer to array of pointers pointing to darray objects
 } darray;
 
-void set_array_obj(obj_t *obj, int index, obj_t *new_obj);
-obj_t *get_array_obj(obj_t *obj, int index);
+void set_array_obj(obj_t *obj, unsigned long index, obj_t *new_obj);
+obj_t *get_array_obj(obj_t *obj, unsigned long index);
 void append(obj_t *obj, obj_t *new_obj);
 
 #endif
